Evaluate new genomes in parallel using OpenMP
In GAIncGA.C two new genomes are added in step. They are evaluated
as part of the replace function and the final call to evaluate the
population.

In order to do the two evaluations in parallel, we call the evaluate
function explicitly to update their score before they are added.